Chatbot-backend-main/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart>=0.0.18
google-generativeai==0.3.2
google-cloud-aiplatform>=1.38.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.0
google-auth>=2.0.0
python-dotenv==1.0.1
PyPDF2==3.0.1
pymupdf==1.23.7
pillow>=10.3.0
pdf2image==1.16.3
requests==2.31.0
beautifulsoup4==4.12.2
aiofiles==24.1.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
boto3==1.37.22
youtube_transcript_api
playwright
yt-dlp
tiktoken
supabase>=2.0.0
pandas>=1.3.0
openpyxl>=3.0.0
xlrd>=2.0.0
aiohttp>=3.8.0
python-docx>=0.8.11
# 軽量RAGシステム用
bm25s>=0.2.0
scikit-learn>=1.0.0
numpy>=1.21.0
# エンベディングシステム用
google-genai>=0.2.2
pgvector>=0.2.0